@@ -30,9 +30,7 @@ public class SerializedObjectConverter extends TypeConverter {
         try {
             final boolean useCompression = !f.getAnnotation(Serialized.class).disableCompression();
             return Serializer.deserialize(fromDBObject, useCompression);
-        } catch (IOException e) {
-            throw new MappingException("While deserializing to " + f.getFullName(), e);
-        } catch (ClassNotFoundException e) {
+        } catch (IOException | ClassNotFoundException e) {
             throw new MappingException("While deserializing to " + f.getFullName(), e);
         }
     }
