@@ -11,6 +11,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static java.util.stream.Collectors.toList;
 import static org.mongodb.morphia.query.QueryValidator.validateQuery;
 
 
@@ -21,7 +22,7 @@ import static org.mongodb.morphia.query.QueryValidator.validateQuery;
 public class UpdateOpsImpl<T> implements UpdateOperations<T> {
     private final Mapper mapper;
     private final Class<T> clazz;
-    private Map<String, Map<String, Object>> ops = new HashMap<String, Map<String, Object>>();
+    private Map<String, Map<String, Object>> ops = new HashMap<>();
     private boolean validateNames = true;
     private boolean validateTypes = true;
     private boolean isolated;
@@ -229,7 +230,7 @@ public class UpdateOpsImpl<T> implements UpdateOperations<T> {
         final String opString = op.val();
 
         if (!ops.containsKey(opString)) {
-            ops.put(opString, new HashMap<String, Object>());
+            ops.put(opString, new HashMap<>());
         }
         ops.get(opString).put(sb.toString(), val);
     }
@@ -240,10 +241,10 @@ public class UpdateOpsImpl<T> implements UpdateOperations<T> {
     }
 
     protected List<Object> toDBObjList(final MappedField mf, final List<?> values) {
-        final List<Object> list = new ArrayList<Object>(values.size());
-        for (final Object obj : values) {
-            list.add(mapper.toMongoObject(mf, null, obj));
-        }
+        final List<Object> list = new ArrayList<>(values.size());
+        list.addAll(values.stream()
+                          .map(obj -> mapper.toMongoObject(mf, null, obj))
+                          .collect(toList()));
 
         return list;
     }
