@@ -1,11 +1,11 @@
 /**
  * Copyright (C) 2010 Olafur Gauti Gudmundsson
- * <p/>
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may
  * obtain a copy of the License at
- * <p/>
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- * <p/>
+ * <p>
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
@@ -25,7 +25,6 @@ import org.mongodb.morphia.mapping.MappedField;
 import org.mongodb.morphia.mapping.Mapper;
 
 import java.lang.annotation.Annotation;
-import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
@@ -84,13 +83,13 @@ public class ExternalMapperExtTest extends TestBase {
             final MappedClass destMC = mapper.getMappedClass(destClass);
             final MappedClass sourceMC = mapper.getMappedClass(sourceClass);
             //copy the class level annotations
-            for (final Entry<Class<? extends Annotation>, List<Annotation>> e : sourceMC.getRelevantAnnotations().entrySet()) {
-                if (e.getValue() != null && !e.getValue().isEmpty()) {
-                    for (final Annotation ann : e.getValue()) {
-                        destMC.addAnnotation(e.getKey(), ann);
-                    }
-                }
-            }
+            sourceMC.getRelevantAnnotations().entrySet().stream()
+                    .filter(e -> e.getValue() != null && !e.getValue().isEmpty())
+                    .forEach(e -> {
+                        for (final Annotation ann : e.getValue()) {
+                            destMC.addAnnotation(e.getKey(), ann);
+                        }
+                    });
             //copy the fields.
             for (final MappedField mf : sourceMC.getPersistenceFields()) {
                 final Map<Class<? extends Annotation>, Annotation> annMap = mf.getAnnotations();
