@@ -45,9 +45,7 @@ public class TestAsListPerf extends TestBase {
             mongoThreads.add(new MongoQueryThread(mongoQueryThreadsResult, nbOfAddresses));
         }
         final ExecutorService mongoPool = Executors.newFixedThreadPool(threadPool);
-        for (final MongoQueryThread mongoQueryThread : mongoThreads) {
-            mongoPool.execute(mongoQueryThread);
-        }
+        mongoThreads.forEach(mongoPool::execute);
 
         mongoPool.shutdown();
         mongoPool.awaitTermination(30, TimeUnit.SECONDS);
@@ -58,9 +56,7 @@ public class TestAsListPerf extends TestBase {
             morphiaThreads.add(new MorphiaQueryThread(morphiaQueryThreadsResult, nbOfAddresses));
         }
         final ExecutorService morphiaPool = Executors.newFixedThreadPool(threadPool);
-        for (final MorphiaQueryThread thread : morphiaThreads) {
-            morphiaPool.execute(thread);
-        }
+        morphiaThreads.forEach(morphiaPool::execute);
         morphiaPool.shutdown();
         morphiaPool.awaitTermination(30, TimeUnit.SECONDS);
         LOG.debug(format("compareDriverAndMorphiaQueryingMultithreaded (%d queries each) - driver: %4.2f ms/pojo (avg), "
@@ -85,9 +81,7 @@ public class TestAsListPerf extends TestBase {
             morphiaThreads.add(new MorphiaQueryThread(morphiaQueryThreadsResult, nbOfAddresses));
         }
         final ExecutorService morphiaPool = Executors.newFixedThreadPool(threadPool);
-        for (final MorphiaQueryThread thread : morphiaThreads) {
-            morphiaPool.execute(thread);
-        }
+        morphiaThreads.forEach(morphiaPool::execute);
 
         morphiaPool.shutdown();
         morphiaPool.awaitTermination(30, TimeUnit.SECONDS);
@@ -98,9 +92,7 @@ public class TestAsListPerf extends TestBase {
             mongoThreads.add(new MongoQueryThread(mongoQueryThreadsResult, nbOfAddresses));
         }
         final ExecutorService mongoPool = Executors.newFixedThreadPool(threadPool);
-        for (final MongoQueryThread mongoQueryThread : mongoThreads) {
-            mongoPool.execute(mongoQueryThread);
-        }
+        mongoThreads.forEach(mongoPool::execute);
 
         mongoPool.shutdown();
         mongoPool.awaitTermination(30, TimeUnit.SECONDS);
@@ -117,7 +109,7 @@ public class TestAsListPerf extends TestBase {
                                            .sort(new BasicDBObject("name", 1))
                                            .toArray();
         final EntityCache entityCache = new DefaultEntityCache();
-        final List<Address> resultList = new LinkedList<Address>();
+        final List<Address> resultList = new LinkedList<>();
         for (final DBObject dbObject : list) {
             final Address address = getMorphia().fromDBObject(Address.class, dbObject, entityCache);
             resultList.add(address);
@@ -130,14 +122,12 @@ public class TestAsListPerf extends TestBase {
     @Test
     public void driverQueryingMultithreaded() throws InterruptedException {
         final Result mongoQueryThreadsResult = new Result(nbOfTasks);
-        final List<MongoQueryThread> mongoThreads = new ArrayList<MongoQueryThread>(nbOfTasks);
+        final List<MongoQueryThread> mongoThreads = new ArrayList<>(nbOfTasks);
         for (int i = 0; i < nbOfTasks; i++) {
             mongoThreads.add(new MongoQueryThread(mongoQueryThreadsResult, nbOfAddresses));
         }
         final ExecutorService mongoPool = Executors.newFixedThreadPool(threadPool);
-        for (final MongoQueryThread mongoQueryThread : mongoThreads) {
-            mongoPool.execute(mongoQueryThread);
-        }
+        mongoThreads.forEach(mongoPool::execute);
 
         mongoPool.shutdown();
         mongoPool.awaitTermination(30, TimeUnit.SECONDS);
@@ -160,14 +150,12 @@ public class TestAsListPerf extends TestBase {
     @Test
     public void morphiaQueryingMultithreaded() throws InterruptedException {
         final Result morphiaQueryThreadsResult = new Result(nbOfTasks);
-        final List<MorphiaQueryThread> morphiaThreads = new ArrayList<MorphiaQueryThread>(nbOfTasks);
+        final List<MorphiaQueryThread> morphiaThreads = new ArrayList<>(nbOfTasks);
         for (int i = 0; i < nbOfTasks; i++) {
             morphiaThreads.add(new MorphiaQueryThread(morphiaQueryThreadsResult, nbOfAddresses));
         }
         final ExecutorService morphiaPool = Executors.newFixedThreadPool(threadPool);
-        for (final MorphiaQueryThread thread : morphiaThreads) {
-            morphiaPool.execute(thread);
-        }
+        morphiaThreads.forEach(morphiaPool::execute);
         morphiaPool.shutdown();
         morphiaPool.awaitTermination(30, TimeUnit.SECONDS);
 
@@ -198,7 +186,7 @@ public class TestAsListPerf extends TestBase {
         private final Vector<Double> results;
 
         public Result(final int nbOfHits) {
-            results = new Vector<Double>(nbOfHits);
+            results = new Vector<>(nbOfHits);
         }
 
         public double getAverageTime() {
